Add checking for user authorization

+++ src/app/components/comment-like/comment-like.component.ts
@@ -1,6 +1,7 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { IComment } from 'src/app/interfaces/comment';
 import { ICommentLike } from 'src/app/interfaces/commentLike';
+import { AccountService } from 'src/app/services/account.service';
 import { CommentLikeService } from 'src/app/services/comment-like.service';
 
 @Component({
@@ -9,7 +10,7 @@ import { CommentLikeService } from 'src/app/services/comment-like.service';
   styleUrls: ['./comment-like.component.scss']
 })
 export class CommentLikeComponent implements OnInit {
-  constructor(private commentLikeService: CommentLikeService) {}
+  constructor(private commentLikeService: CommentLikeService, private accountService: AccountService) {}
   ngOnInit(): void {
     this.commentLikeService.getByCommentId(this.comment.id).subscribe(res => {
       const existingLikeIndex = res.findIndex(like => like.userId == this.currentUserId);
@@ -22,22 +23,24 @@ export class CommentLikeComponent implements OnInit {
   liked: boolean = false;
 
   toggleLike() {
-    this.liked = !this.liked;
-    const existingLikeIndex = this.comment?.commentLikes?.findIndex(like => like.userId == this.currentUserId);
-    if (existingLikeIndex != undefined && existingLikeIndex >= 0) {
-      const existingLikeId = this.comment?.commentLikes?.[existingLikeIndex]?.id ?? 0;
-      this.commentLikeService.delete(existingLikeId).subscribe();
-      this.comment?.commentLikes?.splice(existingLikeIndex, 1);
-    }
-    else {
-      const newLike: ICommentLike = {
-        id: 0,
-        userId: this.currentUserId,
-        commentId: this.comment?.id || 0
-      };
-      this.commentLikeService.create(newLike).subscribe(res => {
-        this.commentLikeService.getByCommentId(this.comment.id).subscribe(res => this.comment.commentLikes = res)
-      })
+    if (this.accountService.isAuthorized()) {
+      this.liked = !this.liked;
+      const existingLikeIndex = this.comment?.commentLikes?.findIndex(like => like.userId == this.currentUserId);
+      if (existingLikeIndex != undefined && existingLikeIndex >= 0) {
+        const existingLikeId = this.comment?.commentLikes?.[existingLikeIndex]?.id ?? 0;
+        this.commentLikeService.delete(existingLikeId).subscribe();
+        this.comment?.commentLikes?.splice(existingLikeIndex, 1);
+      }
+      else {
+        const newLike: ICommentLike = {
+          id: 0,
+          userId: this.currentUserId,
+          commentId: this.comment?.id || 0
+        };
+        this.commentLikeService.create(newLike).subscribe(res => {
+          this.commentLikeService.getByCommentId(this.comment.id).subscribe(res => this.comment.commentLikes = res)
+        })
+      }
     }
   }
 }
